datasets: Handle converting `int{16,32}` audio data in `VoiceSample`.
We saw `VoiceSample` failing the assert on `float32` audio data when
playing around with the Gradio infer app and submitting an `mp3`
file. We didn't dig deeper into Gradio (I'm sure it's possible to
alter/convert there as well), but it seems potentially useful for
`VoiceSample` to handle `int16` and `int32` audio data on top of what
it already handles.